Revert int->long

+++ python/yt/yson/yson_types.py
@@ -27,9 +27,9 @@ class YsonString(str, YsonType):
     def __repr__(self):
         return self.repr(str)
 
-class YsonInt64(long, YsonType):
+class YsonInt64(int, YsonType):
     def __eq__(self, other):
-        return long(self) == long(other) and YsonType.__eq__(self, other)
+        return int(self) == int(other) and YsonType.__eq__(self, other)
 
     def __hash__(self):
         return self.base_hash(int)
